test updates and 11/1/22 pool test prep

+++ test/riptide_autonomy/tools/TestBufferedSubscriber.cpp
@@ -7,12 +7,12 @@ using namespace std::chrono_literals;
 //test to make sure that no messages are received if none are being sent
 TEST(TestToolTest, test_BufferedSubscriber_noMsgs) {
     BufferedSubscriber<std_msgs::msg::String> bufferedSub(BtTestEnvironment::getBtTestTool(), "test_topic");
-    ASSERT_EQ(bufferedSub.getMessages().size(), (unsigned int) 0);
+    ASSERT_EQ(bufferedSub.getMessages().size(), 0U);
 
     //sleep 1 second
     usleep(1000000);
 
-    ASSERT_EQ(bufferedSub.getMessages().size(), (unsigned int) 0);
+    ASSERT_EQ(bufferedSub.getMessages().size(), 0U);
 }
 
 TEST(TestToolTest, test_BufferedSubscriber_someMsgs) {
@@ -36,10 +36,10 @@ TEST(TestToolTest, test_BufferedSubscriber_someMsgs) {
 
 TEST(TestToolTest, test_BufferedSubscriber_clear_no_msgs) {
     BufferedSubscriber<std_msgs::msg::String> bufferedSub(BtTestEnvironment::getBtTestTool(), "test_topic");
-    ASSERT_EQ(bufferedSub.getMessages().size(), (unsigned int) 0);
+    ASSERT_EQ(bufferedSub.getMessages().size(), 0U);
     
     bufferedSub.clearMessages();
-    ASSERT_EQ(bufferedSub.getMessages().size(), (unsigned int) 0);
+    ASSERT_EQ(bufferedSub.getMessages().size(), 0U);
 }
 
 TEST(TestToolTest, test_BufferedSubscriber_clear_msgs) {
@@ -56,5 +56,5 @@ TEST(TestToolTest, test_BufferedSubscriber_clear_msgs) {
 
     ASSERT_GT(bufferedSub.getMessages().size(), (unsigned int) 7);
     bufferedSub.clearMessages();
-    ASSERT_EQ(bufferedSub.getMessages().size(), (unsigned int) 0);
+    ASSERT_EQ(bufferedSub.getMessages().size(), 0U);
 }
